Added game loding in background thread.
In addition to adding the threadding, I greatly refactored the code.
For the most part, I made it so a Profile (now called SteamProfile)
must have a vadid steamID for it to be created. This changed how
it was being sent in and received from the ProfileReteriever (which
is not working currently). I don't not 100% if I like this but it
would be easy to revert.